application/serializers.py
from rest_framework import serializers
from .models import Post
from users.serializers import UserSerializer


class PostSerializer(serializers.ModelSerializer):
    author = UserSerializer(required=False)

    class Meta:
        model = Post
        fields = ('pk', 'title', 'content', 'like', 'author', 'status', 'created_on')