requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-email-utils==0.1.1
django-rest-email-auth==2.1.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
email-hunter-python==1.1.0
idna==2.9
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
urllib3==1.25.8
